tests/wpt/tests/fetch/compression-dictionary/dictionary-clear-site-data.tentative.https.html

<!DOCTYPE html>
<head>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/compression-dictionary-util.js"></script>
</head>
<body>
<script>

compression_dictionary_promise_test(async (t) => {
  const dict = await (await fetch(kRegisterDictionaryPath)).text();
  // Wait until `available-dictionary` header is available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {}),
      kDefaultDictionaryHashBase64);
  // Clear site data.
  assert_equals(await clearSiteData(/*directive=*/'cache'), 'OK');
  // Check if `available-dictionary` header is not available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {max_retry: 0}),
      '"available-dictionary" header is not available');
}, 'Clear-Site-Data with "cache" directive must unregister dictionary');

compression_dictionary_promise_test(async (t) => {
  const dict = await (await fetch(kRegisterDictionaryPath)).text();
  // Wait until `available-dictionary` header is available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {}),
      kDefaultDictionaryHashBase64);
  // Clear site data.
  assert_equals(await clearSiteData(/*directive=*/'cookies'), 'OK');
  // Check if `available-dictionary` header is not available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {max_retry: 0}),
      '"available-dictionary" header is not available');
}, 'Clear-Site-Data with "cookies" directive must unregister dictionary');

compression_dictionary_promise_test(async (t) => {
  const dict = await (await fetch(kRegisterDictionaryPath)).text();
  // Wait until `available-dictionary` header is available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {}),
      kDefaultDictionaryHashBase64);
  // Clear site data.
  assert_equals(await clearSiteData(/*directive=*/'storage'), 'OK');
  // Check if `available-dictionary` header is not available.
  assert_equals(
      await waitUntilAvailableDictionaryHeader(t, {max_retry: 0}),
      kDefaultDictionaryHashBase64);
}, 'Clear-Site-Data with "storage" directive must not unregister dictionary');

</script>
</body>
